CLean code
