+++ extension/war/src/main/webapp/WEB-INF/conf/social-extension/social/component-plugins-configuration.xml
@@ -99,6 +99,11 @@
       <type>org.exoplatform.social.core.listeners.SocialMembershipListenerImpl</type>
     </component-plugin>
     <component-plugin>
+      <name>social.update.membership.event.listener</name>
+      <set-method>addListenerPlugin</set-method>
+      <type>org.exoplatform.social.core.listeners.UsersProfileGroupsIdsListenerImpl</type>
+    </component-plugin>
+    <component-plugin>
       <name>social.update.profile.event.listener</name>
       <set-method>addListenerPlugin</set-method>
       <type>org.exoplatform.social.core.listeners.SocialUserProfileEventListenerImpl</type>
